# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Queue API
  description: |
    A distributed task queue system using Redis Streams for reliable message delivery.

    Features:
    - Priority-based task processing (critical, high, normal, low)
    - Scheduled tasks for delayed execution
    - Automatic retries with exponential backoff
    - Dead letter queue for failed tasks
    - Real-time WebSocket events
    - Worker management (pause/resume)
    - Rate limiting and backpressure
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task by ID
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Tasks
      summary: Cancel a task
      description: Cancels a pending, scheduled, or retrying task. Cannot cancel running, completed, or failed tasks.
      operationId: cancelTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Task cannot be cancelled in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    taskId:
      name: taskId
      in: path
      required: true
      description: Task identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        payload:
          type: object
          additionalProperties: true
        priority:
          type: string
          enum: [low, normal, high, critical]
        state:
          type: string
          enum: [pending, scheduled, running, completed, failed, retrying, cancelled, dead_letter]
        attempts:
          type: integer
        max_retries:
          type: integer
        error:
          type: string
        result:
          type: object
          additionalProperties: true
        worker_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        scheduled_at:
          type: string
          format: date-time
          description: When the task is scheduled to run (for delayed/scheduled tasks and retrying tasks).
        next_retry_at:
          type: string
          format: date-time
          description: When the next retry attempt is scheduled. Only present when state is `retrying`.
        metadata:
          type: object
          additionalProperties:
            type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: HTTP status text
          example: Bad Request
        message:
          type: string
          description: Detailed error message
          example: task type is required
...